change Snapshot's read/write extensions to SnapshotType's them.
